Day 18 - redis SIM db/requirements.txt
redis==3.5.3
Flask==1.1.2
flask_restful==0.3.9
PyYAML==6.0
